+++ actix-tls/CHANGES.md
@@ -8,6 +8,8 @@
 
 * Enable rustls acceptor service
 
+* Enable native-tls acceptor service
+
 ## [1.0.0-alpha.1] - 2019-12-02
 
 * Split openssl accetor from actix-server package